requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
djangorestframework==3.12.4
fake-useragent==0.1.11
idna==2.10
lxml==4.6.3
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
